# api/openapi/chat.swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "chat/v1/user.proto",
    "version": "version not set"
  },
...
  "consumes": [
    "application/json"
  ],
  "produces": [
    "application/json"
  ],
  "paths": {
...
    "/api/v1/dialogs": {
      "get": {
        "summary": "Собеседники с последним сообщением, сначала самые свежие",
        "operationId": "MessagesService_ListDialogs",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/v1ListDialogsResponse"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/rpcStatus"
            }
          }
        },
        "parameters": [
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "type": "integer",
            "format": "int32"
          }
        ],
        "tags": [
          "MessagesService"
        ]
      }
    },
    "/api/v1/dialogs/{peerId}/messages": {
      "get": {
        "summary": "История диалога от новых к старым",
        "operationId": "MessagesService_ListMessages",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/v1ListMessagesResponse"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/rpcStatus"
            }
          }
        },
        "parameters": [
          {
            "name": "peerId",
            "in": "path",
            "required": true,
            "type": "string",
            "format": "int64"
          },
          {
            "name": "beforeId",
            "in": "query",
            "required": false,
            "type": "string",
            "format": "int64"
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "type": "integer",
            "format": "int32"
          }
        ],
        "tags": [
          "MessagesService"
        ]
      }
    },
    "/api/v1/messages": {
      "post": {
        "operationId": "MessagesService_SendMessage",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/v1SendMessageResponse"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/rpcStatus"
            }
          }
        },
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/v1SendMessageRequest"
            }
          }
        ],
        "tags": [
          "MessagesService"
        ]
      }
    },
...
  },
  "definitions": {
    "protobufAny": {
      "type": "object",
      "properties": {
        "@type": {
          "type": "string"
        }
      },
      "additionalProperties": {}
    },
    "rpcStatus": {
      "type": "object",
      "properties": {
        "code": {
          "type": "integer",
          "format": "int32"
        },
        "message": {
          "type": "string"
        },
        "details": {
          "type": "array",
          "items": {
            "type": "object",
            "$ref": "#/definitions/protobufAny"
          }
        }
      }
    },
    "v1Dialog": {
      "type": "object",
      "properties": {
        "peer": {
          "$ref": "#/definitions/v1User"
        },
        "lastMessage": {
          "$ref": "#/definitions/v1Message"
        }
      }
    },
...
    "v1ListDialogsResponse": {
      "type": "object",
      "properties": {
        "dialogs": {
          "type": "array",
          "items": {
            "type": "object",
            "$ref": "#/definitions/v1Dialog"
          }
        }
      }
    },
    "v1ListMessagesResponse": {
      "type": "object",
      "properties": {
        "messages": {
          "type": "array",
          "items": {
            "type": "object",
            "$ref": "#/definitions/v1Message"
          }
        },
        "nextBeforeId": {
          "type": "string",
          "format": "int64",
          "title": "не задан, если старше сообщений нет"
        }
      }
    },
...
    "v1Message": {
      "type": "object",
      "properties": {
        "id": {
          "type": "string",
          "format": "int64"
        },
        "senderId": {
          "type": "string",
          "format": "int64"
        },
        "recipientId": {
          "type": "string",
          "format": "int64"
        },
        "text": {
          "type": "string"
        },
        "createdAt": {
          "type": "string",
          "format": "date-time"
        }
      }
    },
...
    "v1SendMessageRequest": {
      "type": "object",
      "properties": {
        "recipientId": {
          "type": "string",
          "format": "int64"
        },
        "text": {
          "type": "string"
        }
      }
    },
    "v1SendMessageResponse": {
      "type": "object",
      "properties": {
        "message": {
          "$ref": "#/definitions/v1Message"
        }
      }
    },
    "v1User": {
      "type": "object",
      "properties": {
        "id": {
          "type": "string",
          "format": "int64"
        },
        "login": {
          "type": "string"
        },
        "createdAt": {
          "type": "string",
          "format": "date-time"
        }
      },
      "description": "User — публичное представление пользователя."
    }
  }
}